+++ packages/backend/src/modules/tracked-txs/modules/liveness/indexers/AnomaliesIndexer.ts
@@ -19,8 +19,8 @@ import {
 import { LivenessRepository } from '../repositories/LivenessRepository'
 import {
   LivenessRecordWithConfig,
-  LivenessWithConfigRepository,
-} from '../repositories/LivenessWithConfigRepository'
+  LivenessWithConfigService,
+} from '../services/LivenessWithConfigService'
 import { RunningStatistics } from '../utils/RollingVariance'
 import { Interval, calculateIntervals } from '../utils/calculateIntervals'
 import { getActiveConfigurations } from '../utils/getActiveConfigurations'
@@ -89,7 +89,7 @@ export class AnomaliesIndexer extends ManagedChildIndexer {
         continue
       }
 
-      const livenessWithConfig = new LivenessWithConfigRepository(
+      const livenessWithConfig = new LivenessWithConfigService(
         activeConfigs,
         this.$.livenessRepository,
       )
