+++ packages/backend/src/modules/tracked-txs/modules/liveness/utils/calculateIntervals.ts
@@ -1,4 +1,4 @@
-import { LivenessRecordWithConfig } from '../repositories/LivenessWithConfigRepository'
+import { LivenessRecordWithConfig } from '../services/LivenessWithConfigService'
 
 export type Interval = {
   record: LivenessRecordWithConfig
