+++ packages/backend/src/modules/tracked-txs/modules/liveness/utils/calculateStats.test.ts
@@ -1,6 +1,6 @@
 import { UnixTime } from '@l2beat/shared-pure'
 import { expect, mockObject } from 'earl'
-import { LivenessRecordWithConfig } from '../repositories/LivenessWithConfigRepository'
+import { LivenessRecordWithConfig } from '../services/LivenessWithConfigService'
 import { Interval } from './calculateIntervals'
 import { calculateStats } from './calculateStats'
