+++ packages/backend/src/modules/tracked-txs/modules/liveness/utils/groupByType.test.ts
@@ -1,7 +1,7 @@
 import { UnixTime } from '@l2beat/shared-pure'
 
 import { expect, mockObject } from 'earl'
-import { LivenessRecordWithConfig } from '../repositories/LivenessWithConfigRepository'
+import { LivenessRecordWithConfig } from '../services/LivenessWithConfigService'
 import { groupByType } from './groupByType'
 
 describe(groupByType.name, () => {
